% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depression.R
\docType{data}
\name{depression}
\alias{depression}
\title{Meta-Analytic Depression Psychotherapy Database Dataset}
\format{
An object of class \code{data.frame} with 411 rows and 20 columns.
}
\source{
Meta-Analytic Depression Psychotherapy Database, maintained by the VU University Amsterdam.
}
\usage{
data(depression)
}
\description{
Meta-Analytic Depression Psychotherapy Database Dataset
}
\author{
Pim Cuijpers
}
\keyword{datasets}
