% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{Import Metapsy databases into the R environment}
\usage{
getData(shorthand, version = NULL)
}
\arguments{
\item{shorthand}{\code{character}. The shorthand assigned to the database that
should be downloaded. Shorthands are listed under "Metadata" in the
\href{https://docs.metapsy.org/databases}{database documentation entry}. Alternatively,
all available databases and their shortcodes can be accessed by running \code{\link[=listData]{listData()}}.}

\item{version}{\code{character}. The version number to be downloaded. Default is
\code{NULL}, which downloads the latest version.}
}
\value{
Returns the selected database as a \code{data.frame} object. Metadata is added
via attributes to the object:
\itemize{
\item \code{attr(x,"databaseDoi")} returns the database/"concept" DOI.
\item \code{attr(x,"versionDoi")} returns the database version DOI.
\item \code{attr(x,"publicationDate")} returns the publication date of
the database (version).
\item \code{attr(x,"lastSearch")} returns the last search update.
\item \code{attr(x,"documentation")} returns the URL to the documentation entry for
the database.
}
}
\description{
This function allows to download all publicly available versions of Metapsy
databases into your R environment.
}
\details{
After the databases has been downloaded and saved as an object,
the \code{\link{variableDescription}} function can be used to print a
variable description. The  \code{\link{openDocs}} function opens the
online documentation entry for the database.
}
\examples{
\dontrun{
# List all available databases
listData()

# Get latest version of the 'depression-psyctr' database
d <- getData("depression-psyctr")

# Get version 22.2 of the 'depression-psyctr' database
d <- getData("depression-psyctr")

# Show variable description
variableDescription(d)

# Open online documentation
openDocs(d)}

}
\seealso{
\code{\link{listData}}, \code{\link{variableDescription}},
\code{\link{openDocs}}
}
\author{
Mathias Harrer \email{mathias.h.harrer@gmail.com}
}
