% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openDocs.R
\name{openDocs}
\alias{openDocs}
\title{Open database documentation}
\usage{
openDocs(data)
}
\arguments{
\item{data}{A database \code{data.frame} imported using \code{\link[=getData]{getData()}}.}
}
\value{
Opens the database documentation entry on
\href{https://docs.metapsy.org/databases/}{docs.metapsy.org}.
}
\description{
This function opens the online documentation of an imported database.
}
\examples{
\dontrun{
# List all available databases
listData()

# Get latest version of the 'depression-psyctr' database
d <- getData("depression-psyctr")

# Get version 22.2 of the 'depression-psyctr' database
d <- getData("depression-psyctr")

# Show variable description
variableDescription(d)

# Open online documentation
openDocs(d)
}

}
\seealso{
\code{\link{listData}}, \code{\link{getData}},
\code{\link{variableDescription}},
}
\author{
Mathias Harrer \email{mathias.h.harrer@gmail.com}
}
