% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listData.R
\name{print.listData}
\alias{print.listData}
\title{Print \code{listData} objects}
\usage{
\method{print}{listData}(x, ...)
}
\arguments{
\item{x}{An object of class \code{listData}.}

\item{...}{Additional arguments.}
}
\description{
Prints the available databases returned by \code{\link[=listData]{listData()}}.
}
\author{
Mathias Harrer \email{mathias.h.harrer@gmail.com},
Paula Kuper \email{paula.r.kuper@gmail.com}, Pim Cuijpers \email{p.cuijpers@vu.nl}
}
