% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{print.metapsyDatabase}
\alias{print.metapsyDatabase}
\title{Print \code{metapsyDatabase} objects}
\usage{
\method{print}{metapsyDatabase}(x, first = NULL, last = NULL, ...)
}
\arguments{
\item{x}{A database of class \code{metapsyDatabase} imported using \code{\link[=getData]{getData()}}.}

\item{first}{Number of first rows to be printed.}

\item{last}{Number of last rows to be printed.}

\item{...}{Additional arguments.}
}
\description{
Prints the dataset (\code{data}) if objects returned by \code{\link[=getData]{getData()}} are of class
\code{metapsyDatabase} (i.e. if \code{include.metadata=TRUE}).
}
\author{
Mathias Harrer \email{mathias.h.harrer@gmail.com},
Paula Kuper \email{paula.r.kuper@gmail.com}, Pim Cuijpers \email{p.cuijpers@vu.nl}
}
