% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableDescription.R
\name{print.variableDescription}
\alias{print.variableDescription}
\title{Print variable description}
\usage{
\method{print}{variableDescription}(x, ...)
}
\arguments{
\item{x}{A database \code{data.frame} imported using \code{\link[=getData]{getData()}}.}

\item{...}{Additional arguments.}
}
\description{
Prints the variable description included in the metadata of an imported
Metapsy database.
}
\author{
Mathias Harrer \email{mathias.h.harrer@gmail.com},
Paula Kuper \email{paula.r.kuper@gmail.com}, Pim Cuijpers \email{p.cuijpers@vu.nl}
}
