% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suicide.R
\docType{data}
\name{suicide}
\alias{suicide}
\title{Meta-Analytic Suicide Prevention Database Dataset}
\format{
An object of class \code{escalc} (inherits from \code{data.frame}) with 113 rows and 21 columns.
}
\source{
Meta-Analytic Suicide Prevention Database, maintained by the VU University Amsterdam and 113 Zelfmoordpreventie.
}
\usage{
data(suicide)
}
\description{
Meta-Analytic Suicide Prevention Database Dataset
}
\author{
Wouter van Ballegooijen, Josine Rawee
}
\keyword{datasets}
