% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{updateZenodoAccessToken}
\alias{updateZenodoAccessToken}
\title{Update Zenodo API Access Token}
\usage{
updateZenodoAccessToken(token = NULL)
}
\arguments{
\item{token}{\code{character}. The Zenodo API access token. If \code{NULL}
(default), the function will interactively prompt you to enter the token.
You can create a token \href{https://zenodo.org/account/settings/applications/tokens/new/}{here}.}
}
\value{
Invisibly returns \code{TRUE} if successful, \code{FALSE} otherwise.
}
\description{
This function allows you to update or set your Zenodo API access token.
The token will be saved to your \code{.Renviron} file for permanent storage
and will also be set for the current R session.
}
\details{
The function performs the following actions:
\itemize{
\item If \code{token} is \code{NULL}, prompts you to enter the token
\item Strips any quotes from the token if accidentally included
\item Saves the token to your \code{.Renviron} file (located at \verb{~/.Renviron})
\item Updates the token in your current R session
\item If a token already exists in \code{.Renviron}, it will be updated
}

After updating the token, you may need to restart R/RStudio for the changes
to take full effect, though the token will be available in the current
session immediately.
}
\examples{
\dontrun{
# Update token interactively
updateZenodoAccessToken()

# Update token directly
updateZenodoAccessToken("your_token_here")
}

}
\seealso{
\code{\link{getData}}
}
\author{
Mathias Harrer \email{mathias.h.harrer@gmail.com}
}
