% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableDescription.R
\name{variableDescription}
\alias{variableDescription}
\title{Show variable description for Metapsy database}
\usage{
variableDescription(data)
}
\arguments{
\item{data}{A database \code{data.frame} imported using \code{\link[=getData]{getData()}}.}
}
\value{
Returns a \code{data.frame} with two columns:
\itemize{
\item \code{variable}. The variable name.
\item \code{description}. The variable description,
including factor level explanations.
}
}
\description{
This function prints the variable description included in the metadata
of an imported Metapsy database.
}
\examples{
\dontrun{
# List all available databases
listData()

# Get latest version of the 'depression-psyctr' database
d <- getData("depression-psyctr")

# Get version 22.2 of the 'depression-psyctr' database
d <- getData("depression-psyctr")

# Show variable description
variableDescription(d)

# Open online documentation
openDocs(d)
}

}
\seealso{
\code{\link{listData}}, \code{\link{getData}},
\code{\link{openDocs}}
}
\author{
Mathias Harrer \email{mathias.h.harrer@gmail.com}
}
