% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listData.R
\name{listData}
\alias{listData}
\title{List all available databases}
\usage{
listData()
}
\value{
Returns a \code{data.frame} of all available databases by shorthand,
along with a URL for the database documentation entry.
}
\description{
This function lists all publicly available Metapsy databases. It prints all
database \href{https://docs.metapsy.org/databases/#shorthand}{shorthands}, which can
be used to import data using \code{\link[=getData]{getData()}}.
}
\examples{
\dontrun{
# List all available databases
listData()

# Get latest version of the 'depression-psyctr' database
d <- getData("depression-psyctr")

# Get version 22.2 of the 'depression-psyctr' database
d <- getData("depression-psyctr")

# Show variable description
variableDescription(d)

# Open online documentation
openDocs(d)
}

}
\seealso{
\code{\link{getData}}, \code{\link{variableDescription}},
\code{\link{openDocs}}
}
\author{
Mathias Harrer \email{mathias.h.harrer@gmail.com}
}
